refactor(core): move 'getTransactionsForWalletByPaymentHash' to facade
