+++ core/api/src/services/ledger/facade/index.ts
@@ -1,3 +1,4 @@
+export * from "./get-transactions"
 export * from "./intraledger"
 export * from "./offchain-receive"
 export * from "./offchain-send"
